refactor: :recycle: move PersonalInformation form to it's own component
