+++ src/components/forms/index.ts
@@ -1,0 +1,1 @@
+export * from './PersonalInformation';
