+++ src/types/form-data.ts
@@ -1,0 +1,4 @@
+export interface FormData {
+  [key: string]: any;
+  existingLoans?: Array<{ lenderName: string; loanAmount: string }>;
+}
